// src/shared/api/generated/__swagger_client/models/EmployeeHireRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type EmployeeHireRequest = {
    personId: string;
    organizationId: string;
    orgUnitId?: string;
    personnelNumber: string;
    positionName?: string;
    rankName?: string;
    hireDate: string;
    responsible?: boolean;
};